restructured project into folders: components, models, service and styles.  Also improved answer component to highlight answer on mouse hover and to allow the user to select the highlighted answer without having to be directly over the radio button
