--- src/Welcome.js
@@ -1,12 +1,0 @@
-import React from 'react';
-//simple welcome screen - needs to be personalised next iteration
-function Welcome() {
-  return (
-    <div>
-      <h2>Welcome to the Quiz App!</h2>
-      <p>Select a category to start the quiz.</p>
-    </div>
-  );
-}
-
-export default Welcome;
