+++ src/index.js
@@ -4,11 +4,14 @@ import './index.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { Container } from 'react-bootstrap';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <App />
+    <Container style={{ minHeight: '100vh', backgroundColor: '#65afb4'  }}>
+      <App />
+    </Container>
   </React.StrictMode>
 );
